Add prereqs dialog for extension request
